Count edges in time interval

Count number of edges in time interval

Usage

count_edges_in_interval(x, start = NULL, end = NULL, number = 30)

Arguments

x

object of class networkDynamic

start

start time for the calculation

end

end time for the calculation

number

number of intervals in between start and end

Value

vector with counts

Details

Counts how many edges happen (ie. have their start within an interval) over the time period (start, end). The time period is split into number intervals, each of equal time length. So, if start = 0, end = 10, and number = 2, then the time period is divided into the time intervals from 0 (inclusive) to 5 (exclusive) and from 5 (inclusive) to 10 (exclusive).

count_edges_in_interval counts the number of edges that start in each interval.

count_unique_edges_in_interval counts the number of unique edges that start in each interval, so edges that occur multiple times are only counted once.