Florentine families

Network data of 16 Florentine families, in igraph format.

Usage

data(florentine, package = "snafun")

Format

A list containing two igraph networks and a data.frame with attributes.

Source

http://networkdata.ics.uci.edu/netdata/html/florentine.html

Details

A list containing two networks and a data.frame with attributes.

This is a data set of 1520 marriage and 15 business ties among Renaissance Florentine families. The data is originally from Padgett (1994) via UCINET and stored as an igraph object.

This dataset is a subset of the original data and contains data on sixteen families.

The two relations are business ties (flobusiness, recorded financial ties such as loans, credits and joint partnerships) and marriage alliances (flomarriage).

In addition, there is a data.frame with attributes:

wealth each family's net wealth in 1427 (in thousands of lira)

priorates the number of priorates (seats on the civic council) held between 1282- 1344

totalties the total number of business or marriage ties in the total dataset of 116 families.

The data are symmetrically coded. This makes sense for marital ties, but is unfortunate for the financial ties (which are almost certainly directed in real life).

Substantively, the data include families who were engaged in a struggle for political control of the city of Florence in around 1430. Two factions were dominant in this struggle: one revolved around the infamous Medicis (9), the other around the powerful Strozzis (15).