Plot a dendrogram of the community structure

Plot a hierarchical cluster analysis on a communities object

plot_comm_dendrogram(
  x,
  labels = NULL,
  hang = 0.1,
  axes = TRUE,
  frame_plot = FALSE,
  ann = TRUE,
  main = "Communities dendrogram",
  sub = "",
  xlab = "",
  ylab = "",
  ...
)

Arguments

x

a communities object

labels

A character vector of labels for the leaves of the tree. By default the row names or row numbers of the original data are used. If labels = FALSE no labels at all are plotted.

hang

The fraction of the plot height by which labels should hang below the rest of the plot. A negative value will cause the labels to hang down from 0.

axes, frame_plot, ann

logical flags as in plot.default.

main, sub, xlab, ylab

character strings for title.

...

Further graphical arguments. E.g., cex controls the size of the labels (if plotted) in the same way as text.

Value

nothing, only the plot is shown

Details

Plot the result of one of the community detection algorithms as a nice dendrogram. Except for x, all arguments are used for the plot function.

This works for the results from the extract_comm_walktrap, extract_comm_fastgreedy, and extract_comm_girvan. It (currently) does not work for the results from extract_comm_louvain.