Plot network slices

Plot slices of a networkDynamic network

plot_network_slices(x, number = 9, start = NULL, end = NULL, digits = 3)

Arguments

x

object of class networkDynamic

number

number of intervals in between start and end

start

start time for the calculation

end

end time for the calculation

digits

number of decimals, used to show the time slices in the plot

Value

plot

Details

This function is a variation of filmstrip. That function has the odd behavior of only plotting edges that occur at specific times, rather than over intervals. The current function plots the edges that occur within time intervals.

This function splits the time interval between the starting time (start) and the end time (end), into number equal-sized intervals. The plots include all edges that occur in that interval.

By default, the overall interval covers the period from where the first edges starts to where the last one ends.

Each interval is left-closed and right-open. This means that it can happen that the very last edge is not included, if it starts and ends at the very last time point. This can be solved by setting the end time (a little) higher than the actual end time.