Save internal data

Save and add to internal data

Usage

save_internal_data(
  ...,
  overwrite_vars = FALSE,
  overwrite = FALSE,
  compress = "bzip2",
  version = 2
)

Arguments

...

the names of the objects to be saved (as symbols or character strings).

overwrite_vars

logical, is it OK to replace/overwrite existing objects with new values?

overwrite

By default the function will not overwrite existing files. If you really want to do so, set this to TRUE.

compress

Choose the type of compression used by save(). Should be one of "gzip", "bzip2", or "xz".

version

The serialization format version to use. The default, 2, was the default format from R 1.4.0 to 3.5.3. Version 3 became the default from R 3.6.0 and can only be read by R versions 3.5.0 and higher.

Value

nothing

Details

Saves internal data. There are arguments whether or not existing objects can be safely overwritten and whether the data object can be overwritten.

NOTE: this is experimental, warranty until the doorstep at best.